+++ front_end/package.json
@@ -47,6 +47,7 @@
     "@radix-ui/react-tabs": "^1.1.3",
     "@syncfusion/ej2-react-navigations": "^29.1.38",
     "ai": "^4.3.9",
+    "axios": "^1.8.4",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -64,8 +65,7 @@
     "sonner": "^2.0.3",
     "tailwind-merge": "^3.0.2",
     "tw-animate-css": "^1.2.4",
-    "zod": "^3.24.2",
-    "axios": "^1.7.9"
+    "zod": "^3.24.2"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
